new receiveReplicationOperations() and replicateOperation()
