+++ twmap.cc
@@ -1,4 +1,4 @@
-#include "twmap.hh"
+#include "twmap-wrapper.hh"
 #include <iostream>
 #include <unistd.h>
